# src/main/java/co/edu/uniquindio/sistematriage/dto/request/AsignacionInputDTO.java
package co.edu.uniquindio.sistematriage.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class AsignacionInputDTO {

    @NotBlank
    private String responsableId;

    private String observacion;

}